Optimize DOMUtils::addValueToMultivalAttr implementation
This is barely ever called. But it's another good example for the
kind of optimization to look for. Let's say we want to add a value
that is already there. Nothing should change. Still what the old
code did was an extra implode() and setAttribute() with the
unchanged value from before.

Change-Id: I09e34c6cfcded520c658efc4bd809f4750c12687