Fix broken WTUtils::NON_EXTENSION_ENCAP_REGEXP
This was added via I26182c6 about a year ago. It's apparently
supposed to be a copy of the regex above, without the extension part.

That extension part is the word "Extension" followed by a slash and
some word:

 Extension(/[^\s]+)

This should be removed entirely.

I think this wasn't causing trouble because something line "mw:/foo"
with a slash directly after the "mw:" doesn't exist anywhere. Still
I found it confusing that the regular expression would accept
something like this.

Change-Id: Ie0eca431d7ba53f7aeb93fade360fda8d4c917c0