Foster annotation meta tags like regular content
In the initial version of the code, the annotation meta tags were
handled like the template meta tags by transforming them into comments
before fostering and dealing with them in a later stage of the
post-processing.

This patch removes this special behaviour and lets annotation tags
getting fostered during the TreeBuilder stage like the rest of the
content. This removes some special handling around fostered content in
the AnnotationDOMRangeBuilder stage. The MarkFosteredContent stage now
handles the annotations, including pulling the fostered end tag of
annotations after the table.

This has some impact on tests:
* The "dummyanno getting roped to template top-level because of
  fostering should get the about attribute of the template" test result
  gets modified because the annotation is now wrapped earlier in the
  transclusion. This does not change the round-trip behaviour because
  the wrapping transclusion guarantees that the content is not edited
  in DOM mode anyway.
* An extra paragraph is added in fostered content in "dummyanno in the
  middle of fostered content should extend to the whole range, including
  fostered content" because the annotation tags are moved at a
  different time. This doesn't feel problematic (but might lead to
  visual-diffs for some corner cases).
* We remove the "Selser: add content in a cell of a table where
  dummyanno gets fostered out should keep the annotation around the
  table": this test doesn't make sense since the extended annotation
  range would forbid that editing operation.

Change-Id: Idc99ef873474d23e5487a985c686621cd697259e