Separate and remove DOMUtils::pathToAncestor() helper
This helper method tries to do multiple, slightly different things.
It effectively behaves like 3 different methods, depending on how it
is called.
1. When called with no $ancestor the `pathToRoot` is returned instead.
   The root is part of the result.
2. Literally walking up the tree from $node until $ancestor is found.
   $ancestor is not part of the result. This is in contrast to the
   behavior above.
3. There is no guarantee the provided $ancestor actually is one. The
   method silently returns the `pathToRoot` instead, with no easy way
   to handle this error. An option would be to make it return null in
   this case.

But there is only a single caller. And it looks like this caller can
easily walk up the chain of parentNodes by himself. It would even be
faster because it would typically stop before reaching $ancestor.

This method is not used anywhere else according to CodeSearch.

Change-Id: I4186f622e434b5118e0da4425f7333eb35ea87fa