Remove deprecated PageConfig::getTitle()
We have access to the LinkTarget object now, so we don't need
implementors to provide an implementation of the abstract ::getTitle()
method.

I *would* just change this to a non-abstract method, but due to
"quirks" we can't actually get the "prefixed text" form of the title
without access to a SiteConfig object, since the SiteConfig is needed
to provide localized names for the namespace prefix.  Any place which
has a PageConfig has a SiteConfig as well, so this isn't a problem
in practice, but it does pose a problem for defining a standalone
API here.  Just remove the method.

Change-Id: I0fdaca36dfff0a3358f625b46f7de88751fa06f5