Remove deprecated PageConfig::getTitle()
We have access to the LinkTarget object now, so we don't need
implementors to provide an implementation of the abstract ::getTitle()
method.

I *would* just change this to a non-abstract method, but due to
"quirks" we can't actually get the "prefixed text" form of the title
without access to a SiteConfig object, since the SiteConfig is needed
to provide localized names for the namespace prefix.  Any place which
has a PageConfig has a SiteConfig as well, so this isn't a problem
in practice, but it does pose a problem for defining a standalone
API here.  Just remove the method.

This is a re-play of I0fdaca36dfff0a3358f625b46f7de88751fa06f5
which was initially reverted because we had not fully removed
all uses of ::getTitle() in core.

Depends-On: I490bb38108d0202b43ea2a9b391b2e664e7d2d48
Change-Id: Ie3b8802153e380d25f9fc8a75469ec1c3a1b7015