Sync parserTests with Cite
The html/parsoid section of the following test is changed while sync'ing
to reflect the updated input:

"T202593: Identical dir attributes are pointless, but should not
generate an error"

The only necessary part of the change is the now capitalized value of
the dir in the data-mw.  The rest of the change removes some old
normalization to better reflect current output.

Change-Id: Ib868c75a5740b053494632e9212087b889ad7b73