Support special page transclusions
As a first pass, we call out to the legacy parser for the output of
special pages, which is HTML ($isHTML in Parser.php) and only returned
if asked to render ($this->ot['html'] in Parser.php) as opposed to
template expansion ($this->ot['pre']).

This approach shares the limitations of extension tags that don't have a
native Parsoid implementation, where Parsoid continues to call the
legacy parser for their output.

Move towards doing this "in Parsoid" is in
I73f8a77ef4abaa8fdc9ce294f03aed835b3b03fe.

Bug: T59886
Change-Id: I1545c68c49dc3d7a31cd1b8c6c268a6d0220ce5a