Add self-link processing
This is a first version of self-links handling in Parsoid. It is not
intended to fully work yet. In particular, language converter should be
involved in detecting self-links, and it currently is not.

Self-link handling also involves handling section links, which do need
to stay clickable. We have a difference between the legacy parser
handling and Parsoid, in that [[#link]] is rendered with
href="./Page_title#link" in Parsoid, rather than with href="#link". To
avoid T198652, we add the mw-selflink-fragment class in all these links
so that preview popups do not get displayed. It is somewhat unclear at
this time whether there exist cases where we want to display a popup on
a self-link to a section: the Popups comment around
EXCLUDED_LINK_SELECTORS seems to point in the direction of that
possibility, but this doesn't seem easy to generate from basic wikitext
a priori.

Eventually, this pass aims at being moved to core (and Parsoid
won't handle redlinks or selflinks itself), but implementing it here
unblocks the first read-views deployments for the time being.

Prior to this patch, Parsoid only omitted the tittle attribute on pure
fragment links (on Main page, [[#frag]] vs [[Main#frag]]), which led to
an inconsistency in the output.  To match the legacy parser output, this
patch removes the title attribute on all self-links.

Bug: T69486
Depends-On: I07725155ef5e04eb4346a90c34cbacbd70e88ea6
Change-Id: Ie377ac1b9e781541aed9bce854fdd55e8de52473