Fix selser issue with inserted gallery
The inModifiedContent flag is set when traversing into a node that has
an inserted diff mark, since it applies to the entire subtree.  The DOM
diff'er doesn't litter all the nodes of the subtree with the diff mark
so the flag is tracked in SerializerState.  As in
WikitextSerializer::serializeNodeInternal, the inModifiedContent flag is
an indication not to reuse source and instead go through regular
serialization.

Currently, the check in extApi->getOrigSrc is insufficient since it
fails to consider the inModifiedContent state.

When VE copies a node, it keeps the associated DSR info for possible
source resuse when serializing.  Because getOrigSrc ignored the
inModifiedContent state and the DOM diff'er doesn't mark up any of the
subtree of an inserted node, source was being reused for copied
galleries, despite any edits made subsequent to copying.

Bug: T355282
Change-Id: I8430515775981f0b84bb5ca64997265090f2f886