Add attr to pre ext to parse content as wikitext
Proposed in https://phabricator.wikimedia.org/T353697#9470436

Clever editors have realized that placing an include directive at the
end of an extension tag name results in the legacy parser bailing on the
tag as an extension and falling back to the allow list of literal html
tags, thus changing the format of the content in the pre tag from
plaintext to wikitext.

In this patch, we add an attribute to the pre extension tag that
indicates which content format the editor is requesting, giving first
class support to the above hack, rather than trying to reproduce in in
Parsoid.

Similarly, legacy parser support is in
I8f6e01685ef167e7bbd8fd25262cf9e189d60674

The test suite shows that we achieve fairly consistent behaviour, only
diverging on list interpretation.

An alternative implementation is in
I7540571e173c63623a0e91945356236e2a411ac5
which disables a tag as an extension in general and falls back to a
literal html tag.

Bug: T353697
Change-Id: I0d665ae4bfca72a417497bb3b4073c18fb65167a