Remove tracking of needTransclusionShadow in TreeBuilderStage
The needTransclusionShadow state was introduced so that the transclusion
meta could be inserted only once after dispatching characters for
multiple string tokens.  However, since 49a5738, we're buffering the
characters instead of dispatching them, so the needTransclusionShadow
state can just be checked at the time the buffer is cleared.

This is how things worked back in c7403da when the this buffering was
introduced the first time.

Another version of this patch was reverted in dd40bd41 because the
leniency introduced to omit the transclusion meta if the entire buffer
was whitespace, rather than just newlines, introduced a regression to be
investigated separately.

Change-Id: I779c7aa0080639463f2628f872b50880747e5cdc