Match HTML spec where transclusion shadows aren't needed
This change was introduced in 063f5d8 but reverted in dd40bd4 because a
regression was noted roundtripping ukwiki:Військові_звання_України.
Test cases are added here that highlight the issue from that page.

In summary, there's an edge case bug in template wrapping.  The bug is
hidden in some cases because of the side-effect of incorrectly adding
transclusion shadows which swallow the foster box.  But, we should not
be emitting transclusion shadows for any whitespace char really because
the html spec says those chars won't be fostered out.  This patch fixes
that and leaves the template wraping bug exposed in all cases.

To explain the test, content that is not template generated gets
fostered however template wrapping includes that content in its parts so
we get duplicated content when roundtripping.  The tests fail wt2wt and
are added to the knowFailures list.

The difference before and after the change in this patch is highlighted
by the second test which passes wt2wt before this change, though only
fortuitously.  Because there's non-newline whitespace character, a
shadow transclusion meta gets emitted (incorrectly) and ends up in the
foster box, which induces the foster box to be transclusion wrapped even
though no transclusion content was fostered.  The entire table ends up
transclusion marked so the content in the foster box gets dropped after
roundtripping.

Ideally, template wrapping should skip the tbody and place the
encapsulation marking on the table element.  And if a table element is
transclusion marked, it needs to wrap its associated foster content as
well.  That all happens naturally when fostered content is detected in
the foster box.

Follows-Up: I56311c835d287a3125e43f7d20b4ca225108910e
Change-Id: I43925677d073b7df868f3b6bcd9b4ff513d13f5c