Remove bogus comment in storeInPageBundle
The comment, added in af6fec1, was intended to avoid logging that the
page contained a duplicate id unnecessarily.  So, rather benign.

However, it was bogus from the beginning.  When generating a uid, it's
confirmed that the id doesn't exist in the page to begin with.  There's
no need to protect mw ids in the source.

Bug: T358588
Change-Id: Ie96f043b3de2a7942bf088e65c6f416617a5ffac