Combine the anchor processing from WrapSectionState and Headings
These processes were doing essentially the same thing but reaching
slightly different results in edge cases.

WrapSectionState benefits from knowing about Headings::dedupeHeadingIds
and when reusedId is set.

Section metadata for heading is generated in the Headings pass and
stored as temporary data until the WrapSectionsState pass.

Headings::processHeadingContent is used instead of
Headings::textContentOf, which is dropped.  These methods were somewhat
analogous.  However, the mw:LanguageVariant handling of the latter is
introduced to the former.

The Headings::genAnchors pass is reordered before section wrapping in
DOMPostProcessor so that it happens after DisplaySpaces are run.  The
test "TOC with french spacing (T324763)" wants it that way.  It's also
an indication that these passes are related could be combined if section
wrapping wasn't optional.

Bug: T358401
Change-Id: Icbb3b0fb27cc47b764a6a591f647057608424450