Look behind tables for fostered content when linting
Rather than trying to find the associated table for fostered content and
all the mechanics of skipping the content in between, instead, use a
look behind from tables for fostered content.

Simplifies the implementation and unifies the linting method with the
others as void and rearrangeable.

A follow up could be written change logWikitextFixups and the extension
lintHandler to not be able to return a nextNode if that's desirable.

Change-Id: I19213c6057c14777368bf122ac60adafad6c84c0