[CMC] Remove ::setPageProperty() in favor of its more-specific replacements
New code should use ::setUnsortedPageProperty() or ::setNumericPageProperty()
instead of the old ::setPageProperty() method.  This avoids a number of
gotchas from the old interface.

Most callers of ::setPageProperty() in core have already been deprecated
with warnings in MW 1.43; the only permissed uses are those where the
call is identical to ::setUnsortedPageProperty(), aka string values.

Change-Id: Ib5ab471fb7117589c96df339f9e311d952a4d1d8