Add missing-image-alt-text lint
Optional lint check for images with missing alt text to help track
accessibility issues in image markup.

A match on this lint does not necessarily indicate an error, as
alt text is sometimes deliberately left empty, but the marker allows
checking the usage in context.

This goes by presence of the alt attribute in output, so implicit alt
text via caption field on an inline image counts as providing alt text,
as does setting an explicitly empty alt= field which should be used
sparingly.

This is meant to be available for tracking tools, with priority=none
to hide it from the main view on Special:LintErrors.

Added test cases for several ways alt text makes it into links.

Bug: T344378
Change-Id: I53b19dc98e783f35afa497b4258c13731171d9d9