Suppress missing-image-alt-text lint on aria-hidden or role=presentation
Images in a subtree containing aria-hidden=true, role=presentation, or
role=none, will have their accessbility info suppressed and it is safe to
not save a lint marker for them.

This should fix false positives on user-provided HTML using a wrapper like
<div aria-hidden=true role=presentation>, as well as any machine-generated
interior HTML using <img> that's marked up with role=presentation
on the actual element.

Note that images inside an area with a display:none style will not be
suppressed; for one you can't tell for sure without knowing the stylesheet,
and for another we expect most such uses to be for fold-out stuff that
could become visible in later JavaScript interactivity.

Bug: T344378
Change-Id: I6df256366c290cba197ab52319b277e22a0cc3e1