Fix regression in Kartographer caused by the emergence of a `null` dsr
This is a fix to a regression caused by
I3cdb8350858219ed9721b22fd6080e7c1168ba10: previously when we loaded
data-parsoid from JSON, we checked to see if certain fields were null
and left them unset if so.

In the "new" version that test was omitted from
DataParsoid::newFromJsonArray() and so `"dsr":null` appeared in some
test case output and caused Kartographer tests to fail.  These tests
were/are pretty brittle, depending on an exact string match of
generated HTML, but the crux was that they invoked Parsoid with the
'clearDSROffsets' flag, which (ab)used ContentUtils::shiftDSR to set
the 'dsr' field to null.  Because this was in an "embedded attribute",
the generated HTML with "dsr":null was reparsed and reserialized,
which served to clean up and remove the `null`.

A follow up will remove the `null` values on *storing/serialization*,
not just on *loading/deserializtion* as well, which is where it more
logically belongs.  This didn't occur previous to the
I3cdb8350858219ed9721b22fd6080e7c1168ba10 path, however, and so could
potentially cause further regressions.  As such it has been split into
a follow up: I7f73704bf46f2290860bc4d8bf7a1d507d1c6993.

Followup-To: I3cdb8350858219ed9721b22fd6080e7c1168ba10
Change-Id: I6cdaf0867ab021000c538b72ff211cdd59210abb