Add a --wt2lint option to bin/parse.php
This results in a slight change of behaviour to what the --linting
option previous outputted.

Before, --linting would both enable the linter pass while parsing and
call DataAccess::logLinterData.  In the non-integrated modes, logging
linter data called error_log on the lints array.  But, in integrated
mode, it would attempt to queue up the lints to add to the integrated
MediaWiki linter table, which generally isn't what you're trying to do
with bin/parse.php

Now, --linting simply enables the linter pass, which is useful for some
debugging and performance work, when you aren't particular concerned
about the output of the lints.

A new --logLinterData is introduced that works how the old --linting
did, and enables --linting by default.  The name also matches the Env
variable with the functionality.

And a --wt2lint option is added that will just output the lints array in
any of the modes, as expected.  It enabled --linting by default.

Bug: T279831
Change-Id: I3c40373625d2ea6f8168f8f263a9b1e385a2dca6