Reset section numbering per template
Section numbering is supposed to correspond to the section of the
template article, not a global number.  The headingIndex wasn't being
reset on pipeline reuse.

The failing test is added to demonstrate that Parsoid lacks the
necessary insight to determine if the section is from a template or one
nested below it.

Bug: T368095
Change-Id: Ifb8d3efa88557a6abdf0e52401eec26cd7e5961c