Remove meaningless digit transform tables from decimal to decimal
This goes along with Ic547077 where you can find a longer explanation.
Short version is: This was copied from core's MessagesAry.php and
MessagesArz.php files. But what core does is only a workaround to
undo parts of the "ar" fallback both languages use. To do the same
here we should actually undo what the unwanted fallback to "ar" did.
Here in this codebase the default for languages without specialized
digits is null. See generateCiteLanguageCSS.php.

Change-Id: I81e573f8c5b2801202ae63e05eddf09b6425ae5c