Reset Grammar::prevOffset on pipeline resuse
The call to Grammar::parse will call Grammar::initInternal, which resets
the cache, so there doesn't appear to be a chance of cache reuse.

However, without resetting prevOffset, cache clearing in the current
pipeline use will be disabled until top level blocks are emitted at a
greater offset than the length of wikitext from all previous uses of the
pipeline.  In effect, we aren't getting any of the benefit of releasing
memory from the backtracking cache on reused pipelines.

Change-Id: Idd01c80d5cd5fa3f715214c1896c865d00848299