Remove unreachable code
nested_block_in_table isn't going to break on a single pipe, the
inline_break requires two for the 'table' stop, so we shouldn't ever get
to tds without being at two pipes.

Further, if we did match a single pipe in a row position with that rule,
it seems wrong.  And was wrong when it went in in 2dbbd3f.  For example,

{|
| testing | hmm | ok | no
|}

should be

<table>
<tbody><tr>
<td>hmm | ok | no
</td></tr></tbody></table>

Change-Id: I14c5ab126bf3784f6811ecfd396c57f554592afa