Prefer literal string image option to prefixed image option
In Esperanto, the localized alias for 'img_left' (ie, left-aligned media)
is 'maldekstra', while the localized alias for 'img_width' (ie, specifying
the pixel width of an image) is '$1ra' (as an alternative to the usual
'$1px').

Parsoid was preferring the prefixed '$1ra' match over the literal
'maldekstra' match, causing bogus file image option errors.

Bug: T372935
Change-Id: Ie4d68b39e04b30fd4d12bea9c70c23c7604a5710