Parsoid <title> element should be actual title, not displaytitle
The displaytitle is the contents of an <h1> in the skin, and can contain
HTML markup, which the <title> element can not.  The value of the
<title> element in the <html> has never been formally specified in
the MediaWiki DOM Spec, but using the page title (not displaytitle)
here has some advantages for clients.  The displaytitle is not
exported in the <head> at this time, but is available in the
ParserOutput.

Bug: T324431
Change-Id: Idf8dc214ead9c7b194960abdf04395c6dd6e6608