Introduce TitleValue class
This is a parallel to the core-side MediaWiki\Title\TitleValue class and
is a pure-value-object implementation of the LinkTarget interface, just
like the core-side TitleValue is.  This allows for representation of
titles in places where namespace localization is not (yet) necessary;
like in core, a TitleValue can be up-converted to a "full" Title object
when you need full localization -- in core up-conversion requires access
to MediaWikiServices, while in Parsoid it requires access to a
SiteConfig.

TitleValue should replace most places where titles are being passed
around as prefixed strings, because the allow unambiguous representation
in cases where (for example) there are potential conflicts between
namespace and interwiki prefixes.

Change-Id: I15c9a2a804e7aad6fab3cee6f644d546df7b10ba