Move storeDataAttribs() out of the ParserPipeline
Maintain the document in "prepared and loaded" state longer, with a
databag not serialized attributes.  This will become a
normal/standardized format for manipulation of Parsoid DOM internally.
In the short term it also helps me preserve information in
data-parsoid.tmp for some additional metrics a little further through
the parse process.

Change-Id: I82ab3cd71dccf9cd57510a29728d99300baf99d5