Scale new-style timing metrics by 1000
It turns out that TimingMetric::observeTiming() in core expects a value
*in milliseconds* even though the timing metric by convention ends in
`_seconds`.  It performs the scaling internally; you are expected to
use ::observeTimingSeconds() if you wanted to report a value in seconds.

Despite being confusing, this is actually convenient for us; we can
eliminate a now-unnecessary "measuresTime" boolean that was responsible
for rescaling milliseconds to seconds in some cases.

Discovered when the timing values for the
`mediawiki_Parsoid_wt2html_parse_seconds` metrics were off by a factor
of 1000 compared to counter metrics measuring the same values.

Improved the documentation a bit to hopefully avoid the same mistake in
the future.

Change-Id: Ida5ea93a5c9362e6a27fbb9ee7334e3ddfb4ed18