Finish refactoring PageBundle/DomPageBundle
Move methods from ContentUtils to PageBundle/DomPageBundle to make it
clearer which form of document we're using.

This patch series includes:
* Ic3c09444cef51767629a9f7fac9e79351bb1fc48
* I2a29f5cb2d57bb9778d12f123dfbc1dbc6d27fa7
* Ib2b6ab08784449f67d273753a147ee92575063a1
* Ibffb2daaf817ce41102211bb9668b29a6e59c0c1 (this patch)

In this patch series we introduce DomPageBundle to represent the
"parsed DOM + page bundle properties" case, which allows us to create
proper factory and conversion methods to/from DomPageBundle,
PageBundle, and "single Document" forms. Internally we are always
working with a "prepared and loaded Document", and that conversion
to/from one of the PageBundle/etc output formats happens only in the
main Parsoid entrypoints.

We are more permissive about passing in PageBundle or DomPageBundle
in entrypoints where a Document was permitted, and enforce more
carefully the fact that arguments and return values from entry
points are not "prepared and loaded" but that we do that at the
very start/end so that internal documents are always "prepared
and loaded".  Extension APIs also use "prepared and loaded" DOM,
and this has been added to the documentation.

To enforce appropriate loading/storage of data attributes we've added
a flag in the document DataBag, which is set in
ContentUtils::createAndLoadDocument() and cleared in
ContentUtils::ppToXML() which are the main/recommended ways Documents
are loaded and stored.  (It's a little unfortunate that these names
aren't symmetric.)  However, ::ppToXML() is also used to serialize
fragments at various places; these will be replaced with live
DocumentFragments in I4e8e56ddf8a103e3172b13bde62b4c6478d46e30 but for
the moment mark those callsites with `'fragment' => true` in the
storage options.

The following public properties and methods have changed type, but the
Parsoid Env should not be used outside of Parsoid itself:
* Env::$pageBundle  (now a DomPageBundle instead of a boolean)
* Env::$topLevelDoc (always prepared and loaded)
* Env::setupTopLevelDoc() (always expects a prepared and loaded doc)

The following public methods have been removed:
* PageBundle::toDom()
  - moved to DomPageBundle::toDom() which returns a prepared and
    loaded document by default.
* PageBundle::toHtml()
  - replaced with PageBundle::toInlineAttributeHtml() or
    PageBundle::toSingleDocumentHtml()
* PageBundle::apply(), ::encodeForHeadElement(), ::decodeFromHeadElement()
  - now part of the private implementation in DomPageBundle
* ContentUtils::extractDpAndSerialize()
  - now part of the private implementation in DomPageBundle
* DOMDataUtils::getPageBundle(), ::injectPageBundle(), ::extractPageBundle()
  - now part of the private implementation in DomPageBundle

The following code search shows most of these were not used outside Parsoid:

  https://codesearch.wmcloud.org/search/?q=extractDpAndSerialize%7CinjectPageBundle%7CextractPageBundle%7CencodeForHeadElement%7CdecodeFromHeadElement

However, PageBundle::apply(), PageBundle::toHtml(), PageBundle::toDom(),
and DOMDataUtils::getPageBundle() are harder to code-search, and in fact
PageBundle::apply() is used in MediaWiki core.  Core was patched to avoid
use of these methods in I9e6b924d62ccc3312f5c70989477da1e2f21c86b.

We try to always use DOMUtils::parseHTML() when a "plain" document is
needed, and ContentUtils::prepareAndLoadDocument() when a "prepared and
loaded" document is needed.  Most callers which load data attributes now
use the 'markNew' flag, which helps round-trip cleanly.

Depends-On: I9e6b924d62ccc3312f5c70989477da1e2f21c86b
Change-Id: Ibffb2daaf817ce41102211bb9668b29a6e59c0c1