Use PHP __clone() methods
As an artifact of our port from JavaScript, we created our own ::clone()
methods rather than using the standard PHP `clone` functionality.

The implementation of `clone` in PHP invokes the `__clone` magic method
after a shallow clone of the object has been made; the implementer is
then responsible for manually cloning any properties for which a deep
clone is necessary.

Probably more of the uses of Utils::clone() can/should be converted like
this.

This patch makes it easier to add non-scalar data to (eg) DataMw, by
clarifying that it is the responsibility of each class to do its own
deep clone by invoking `clone $this->someProperty`, and delegating
to the class represented by `someProperty` to further clone its
own non-scalar components.

Change-Id: I623d9ae29818cb0a7397c85ceb2d99c51a7b8dd8