Increase use of PageBundle/DomPageBundle methods and ContentUtils
Instead of manually injecting/applying page bundles, use the methods
defined in PageBundle/DomPageBundle -- and change the type of the page
bundle attached to the DataBag to a DomPageBundle to associate it
properly with the top level doc.  (This is a temporary change, as a
future patch will remove the page bundle from the DataBag entirely,
which will also resolve the circular reference created here.)

Try to use ContentUtils methods for a "prepared and loaded" document
and DOMUtils::parseHTML for a "plain" document, and avoid the use of
"prepared but not loaded" documents where possible.  This patch
does the low-hanging fruit, fully shifting the codebase to using
these methods will be done in the follow-up,
Ibffb2daaf817ce41102211bb9668b29a6e59c0c1.

Improve documentation about the preparation/loaded state of the
document at various entrypoints.  In general Parsoid should always be
using "prepared and loaded" documents internally, with the exception
of the main Parsoid entrypoint (which uses plain documents) and
stand-alone libraries like DOMCompat (which should work correctly on
both).  For testing, use "prepared and loaded" documents unless
specifically testing an entrypoint or library which should work
on plain documents.

Change-Id: I2a29f5cb2d57bb9778d12f123dfbc1dbc6d27fa7