Introduce fragment types and API
PFragment is a type to represent a MediaWiki Content fragment.  The
"P" stands for "Parsoid", "Parser", "Page", or "MediaWiki Content",
whichever you like best.

The PFragment type is a generalization of the "strip state" method
used by MediaWiki to represent combinations of wikitext and HTML as a
"wikitext string with strip markers".  Our proposed new fragment
generator API for transclusions will pass arguments and return results
as PFragments, allowing interoperability with transclusions which
return wikitext and HTML.

This patch establishes the PFragment type and wires it into the
parser pipeline to allow DataAccess::preprocessWikitext() to return
a PFragment instead of a plain string in the case where strip markers
are embedded in the preprocessed result.  A follow-on patch will
introduce the fragment generator API using PFragment types.

Bug: T374616
Bug: T279094
Change-Id: I0d8e85b01435f37b8bccab3431c9807d46989ae6