Remove sealed fragment from cloned attributes
Instead of the original, which is less confusing.  Note that the updated
tests mean that the Cite extension tests will also need updating, or
dropped because it's already covered.

Note that if we were to remove sealed fragments as described in
I27cb88b6467d4ccef58b909dfc449ba47c0b936b, we'd need to replace this
special cloning for sealed fragments.  An idea would be to set a flag on
nodes cloned while treebuilding, and then propagate that information
when processing embedded attributes so that handlers would know to drop
the cloned fragment.

Follows-Up: I623d9ae29818cb0a7397c85ceb2d99c51a7b8dd8
Depends-On: Ibf8656ab344ed79b3740f3b98e96ff4353375b62
Change-Id: I9a2f75ea137a8a10842191ad0ea3e92eab41202c