String gluing across include directive boundary
The idea here is that, in cases where we know content is going to be
left in place, allow tokenizing to continue by just emitting a start or
end tag, rather than tokenizing in a separate context and then stitching
into the stream.

While this works, it can mess with the TokenCollector include passes
that expect balanced open/close tags in the token stream where, in these
token gluing cases, start/end tags may be nested in other tokens'
attributes.  This suggests eliminating the token collection passes and
moving that logic to the tokenizer.

The known failures changes investigated are because of includes in
attribute position that no longer hide their contents from preceding
grammar rules (which benefits removing the sol_il rule discussed below)
and thus the content and end tags can parse as individual attributes,
which roundtrip to being space separated.  While selser will no doubt
help here, some care might be needed to ensure content isn't dropped in
editing contexts.  A few tests are adjusted to reflect the normalization
that occurs from the space separation of attributes so that they don't
needlessly fail roundtripping.

Remove sol_il rule parameter introduced in 195971b02, it and the code
using it are no longer necessary since include content is tokenized in
place and changes in sol won't be hidden.

Bug: T373776
Change-Id: Ied9ed72be7bbff6d7f323109bc0043f1aa0f4b70