Expand template target for templatearg as well
TemplateHandler::expandTemplate checks if the template target contains a
template token and expands the target if so.  But if the target only
contains a templatearg, it wasn't being expanded.  The check should
include templatearg as well.

The uses of TokenUtils::isTemplateToken were audited and all seemed like
places to check for templatearg too.

The updated parsertest is because, previously, Parsoid treated top-level
"2{{{3}}}" as a valid template target because it ignored template arg in
template target position and treated the surrounding construct as a
transclusion.  Now, Parsoid recognizes "2{{{3}}}" as an invalid target
and doesn't treat the surrounding construct as a transclusion.

Change-Id: Ia81db757b224c7474364c157e8aa64c301d69c77