Remove native special page transclusion implementation
The native implementation only ever supported top level special page
transclusion.  Now that we have PFragments, let the core preprocessor
return the html as a fragment, even for nested transclusions.

The output of the "Special:Search page linking" changes because of a
difference in fallback behaviour when a template doesn't exist, the
preprocessor creates a wikilink to the special page.  Previously, the
native implementation called out for the legacy parser html so if the
special page didn't exist, the legacy parser would parse the link to
html, and that's what would be included.  Whereas, with PFragment, if
the special page doesn't exist, we don't have an html fragment in the
expansion, only wikitext with a wikilink, which Parsoid then parses.

Bug: T356718
Depends-On: Ie50308bde7212cc19d6fe6273ae36e79ae5f94c3
Depends-On: I04afdb5eeba2b5b748347c3687049b1540d80e2c
Change-Id: I79e66fbbe8f60f12cc0f60ba7a6f94063aa5401d