Fix fostered flag confusion between pipelines
There are two fixes in this patch.

If the fostered flag is set on the first node in a dom fragment, it
shouldn't be transferred to the representative placeholder node since it
doesn't apply to outer pipeline and can confuse the fragment unpacking
pass into thinking that the placeholder was fostered and thus the
fragment nodes all need the flag set. Removing this fix will erroneously
mark nodes in the added regression test as fostered.

The second fix recognizes that fragment content need not all be element
nodes so span wrapping is required if the fostered flag is to be
preserved.  Without this fix, the regression test crashes as in the
task.

Bug: T384644
Change-Id: If2d9962137d44c9c362938cd2ea7b0c63e3777ee