Better handle expanded attrs when converting in table fixups
Prior to this change, the code assumed all the content would be stored
in the first attribute's key and didn't account for the typeof or
data-mw.

Rather than trying to reassemble the output from data-mw, abandon that
approach entirely and treat this case like the rest, do a reparse and
unset data associated with the expanded attr.

As an example of how fraught trying to reconstitute the output from the
set attributes, take

<div class="123" style="<nowiki>test</nowiki>">123</div>

which parses to

<div class="123" style="test" about="#mwt2" typeof="mw:ExpandedAttrs" data-parsoid='{"stx":"html","a":{"style":"test"},"sa":{"style":"&lt;nowiki>test&lt;/nowiki>"},"dsr":[0,56,47,6]}' data-mw='{"attribs":[[{"txt":"style"},{"html":"&lt;span typeof=\"mw:Nowiki\" data-parsoid=&apos;{\"dsr\":[24,45,8,9]}&apos;>test&lt;/span>"}]]}'>123</div>

Not all attributes are in the data-mw (class is only set on the node)
and it's a bit unclear what the order to process the attributes on the
node.

Bug: T384739
Change-Id: I80d904476f2998daaa37ffa87c5576a3cb56afa6