PFragment support: expand parser function result to tokens not DOM
In order to allow tokens from a parser function expansion to be
integrated into the surrounding context, expand parser function
fragments to tokens, not to DOM, the same way template
transclusions are evaluated.

Parser function expansions are typically composite DOM fragments.
The atomic DOM fragments within these (usually strip markers in
the original wikitext on the core side) can be evaluated to DOM
since they don't have any wikitext content.

The associated parserTests should have been in
`html/parsoid+integrated` mode from creation, as this parser
function support relies on the legacy preprocessor and is not
expected to work with the partial parser function support in the
Parsoid-native template expansion code.  Rather than change this
on the core side and have to deal with another sync, I've just
changed the test mode to integrated-only in this patch.

One test fails in wt2wt mode due to a shift in where the <nowiki>
tags are applied in html2wt mode;
I7c17110e8df457dbc312673577ec1b411195eba2 contains a fix for this
but it might not be worth addressing.

Depends-On: I07bacdb4bbe90728d2faa207c19fb92ad0e4a257
Change-Id: Ifd9068d5ae5016bba5a56490cbced42c799f7090