Process white space tokens after <p> tags out of order in Remex
If there is a misnested formatting element in a <p> tag, they are closed
and re-opened around character runs until they are effectively close
(there's a stack of Active Formatting Elements in Remex taking care of
that.)

But, whitespace (including the newlines that we keep between paragraphs) is
counted as character runs, which induces opening/closing of AFEs for these
elements, which can lead to rendering issues such as the ones reported
in this ticket.

In this patch, we look at whitespace after paragraphs: if a paragraph is
followed by whitespace, and if a formatting element is active, we process
the whitespace inside the paragraph rather than outside. It avoids the
AFE reconstruction to be triggered on whitespace (including newlines)
between paragraphs.

Bug: T368720
Depends-On: I864e079fe6191afc98b95208eb6553cac0018cbb
Change-Id: I7665fee2b25ceff693427771ed1ef580bf8e2965