Process new line tokens after <p> tags out of order in Remex
If there is a misnested formatting element in a <p> tag, they are closed
and re-opened around character runs until they are effectively close
(there's a stack of Active Formatting Elements in Remex taking care of
that.)

But, whitespace (including the newlines that we keep between paragraphs) is
counted as character runs, which induces opening/closing of AFEs for these
elements, which can lead to rendering issues such as the ones reported
in this ticket.

In this patch, we look at whitespace after paragraphs: if a paragraph is
followed exactly two new lines, and if a formatting element is active,
we process the new lines inside the paragraph rather than outside.
It avoids the AFE reconstruction to be triggered on new lines between
paragraphs.

Note that we restrict ourselves to "exactly two newlines" because larger
amounts of newlines and/or whitespace in general have proven to be
problematic for wt2wt. As it is, wt2wt is not perfect and, in
particular, fails on the following pattern:

<code>
a

</code>

where the closing </code> disappears (the reason for that is that it is
inserted inside the paragraph with autoInserted: true, because at that
point it is, and the closing one is discarded as it has been closed
already).

Bug: T368720
Change-Id: I231e42aa1bae988e5fcc41d6946d96e7dc603998